fixed the algorithm that split the cation and the anion
fixed a hack that determined the cation and the anion based on the
length of the string.  this was a fatal error because it would make it
so compounds like kBr not identified as binary ionic compounds, this was
because it would split the string so that the cation would = "kB" and
the second string would be "r ".  this time I am seperating the strings
using regex.